Reworked endpoints and services.
