+++ src/api/routes/contest/submissions.ts
@@ -11,9 +11,14 @@ import {
   isString,
   required,
   isNumber,
+  minNumber,
+  isArray,
 } from '../../../../deps.ts';
 import { INewSubmission } from '../../../interfaces/submissions.ts';
 import SubmissionModel from '../../../models/submissions.ts';
+import Top3Model from '../../../models/top3.ts';
+import WinnerModel from '../../../models/winners.ts';
+import AdminService from '../../../services/admin.ts';
 import SubmissionService from '../../../services/submission.ts';
 import authHandler from '../../middlewares/authHandler.ts';
 import upload from '../../middlewares/upload.ts';
@@ -24,9 +29,10 @@ const route = Router();
 export default (app: IRouter) => {
   const logger: log.Logger = serviceCollection.get('logger');
   const subServiceInstance = serviceCollection.get(SubmissionService);
-  const subModelInstance = serviceCollection.get(SubmissionModel);
+  const adminServiceInstance = serviceCollection.get(AdminService);
   app.use(['/submit', '/submission', '/submissions'], route);
 
+  // POST /
   route.post(
     '/',
     authHandler({ adminOnly: false, authRequired: true }),
@@ -70,31 +76,96 @@ export default (app: IRouter) => {
     }
   );
 
+  // GET / <- can be filtered - TODO - admin only!
   route.get(
-    '/recent',
-    authHandler({ authRequired: true, adminOnly: false }),
+    '/',
+    authHandler({ adminOnly: false, authRequired: false }),
     async (req: Request, res: Response, next: NextFunction) => {
       try {
-        const subs = await subModelInstance.get(
-          { userId: req.body.userInfo.id },
-          {
-            limit: 6,
-            orderBy: 'created_at',
-            order: 'DESC',
-          }
-        );
+        // TODO read query params into a generic query!
+        const submissionModelInstance = serviceCollection.get(SubmissionModel);
+        const top10 = await submissionModelInstance.getTodaysTop10();
 
-        const subItems = await Promise.all(
-          subs.map((s) =>
-            subServiceInstance.retrieveSubItem(s, req.body.userInfo.codename)
-          )
-        );
-
-        res.setStatus(200).json(subItems);
+        res.setStatus(200).json(top10);
       } catch (err) {
         logger.error(err);
         throw err;
       }
+    }
+  );
+
+  // GET /top
+  route.get('/top', async (req: Request, res: Response, next: NextFunction) => {
+    try {
+      const top3ModelInstance = serviceCollection.get(Top3Model);
+      const subs = await top3ModelInstance.get();
+
+      res.setStatus(200).json(subs);
+    } catch (err) {
+      logger.error(err);
+      throw err;
+    }
+  });
+
+  // POST /top
+  route.post(
+    '/top',
+    validate({
+      ids: validateArray(true, [minNumber(1)], {
+        minLength: 3,
+        maxLength: 3,
+      }),
+    }),
+    async (req: Request, res: Response, next: NextFunction) => {
+      // TODO - move this into submission service!
+      const top3 = await adminServiceInstance.setTop3(req.body.ids);
+
+      res.setStatus(201).json({ top3, message: 'Top 3 successfully set!' });
+    }
+  );
+
+  // GET /winner
+  route.get(
+    '/winner',
+    async (req: Request, res: Response, next: NextFunction) => {
+      try {
+        const winnerModelInstance = serviceCollection.get(WinnerModel);
+        const winner = await winnerModelInstance.get();
+
+        res.setStatus(200).json(winner);
+      } catch (err) {
+        logger.error(err);
+        throw err;
+      }
+    }
+  );
+
+  // GET /flags
+  route.get(
+    '/flags',
+    validate({ submissionId: [required] }, 'query'),
+    async (req: Request, res: Response, next: NextFunction) => {
+      const flags = await adminServiceInstance.getFlagsBySubId(
+        parseInt(req.query.submissionId)
+      );
+      res.setStatus(200).json({ flags, message: 'Received flags' });
+    }
+  );
+
+  // POST /flags
+  route.post(
+    '/flags',
+    validate({ submissionId: [required] }, 'query'),
+    validate({ flags: [isArray] }, 'body'),
+    async (req: Request, res: Response, next: NextFunction) => {
+      const flags = await adminServiceInstance.flagSubmission(
+        req.query.submissionId,
+        req.body.flags
+      );
+
+      res
+        .setStatus(201)
+        .json({ flags, message: 'Successfully flagged submission' });
     }
   );
